More imx351uc cross fixes

+++ cross-compilers/compilers_linux-imx351uc/build.sh
@@ -1,4 +1,4 @@
-echo $(pwd)
+CHOST="arm-unknown-linux-uclibcgnueabi"
 mkdir -p .build/src
 mkdir -p .build/tarballs
 
@@ -11,30 +11,23 @@ elif [ "$OSTYPE" == "msys" ]; then
   BUILD_NCPUS=${NUMBER_OF_PROCESSORS}
 fi
 
-cp ${RECIPE_DIR}/config .config
-
-sed -i.bak "s|@PREFIX@|${PREFIX}|g" .config
-rm .config.bak
-sed -i.bak "s|@BUILD_TOP@|${PWD}|g" .config
-rm .config.bak
-sed -i.bak "s|CT_PARALLEL_JOBS=4|CT_PARALLEL_JOBS=${BUILD_NCPUS}|g" .config
-rm .config.bak
-
 # If dirty is unset or the g++ binary doesn't exist yet, then run ct-ng
 # TODO :: Change && to ||, this is to prevent rebuilds when I forget to add --dirty.
 #if [[ -z ${DIRTY} && ! -e "${PREFIX}"/bin/arm-unknown-linux-uclibcgnueabi-g++ ]]; then
-if [[ ! -e "${PREFIX}"/bin/arm-unknown-linux-uclibcgnueabi-g++ ]]; then
-   ct-ng build
+if [[ ! -e "${SRC_DIR}/gcc_built/bin/${CHOST}-gcc" ]]; then
+    mv config .config
+    # sed -i.bak "s|@BUILD_TOP@|${PWD}|g" .config
+    # rm .config.bak
+    sed -i.bak "s|CT_PARALLEL_JOBS=4|CT_PARALLEL_JOBS=${BUILD_NCPUS}|g" .config
+    rm .config.bak
+    # building oldconfig will ensure that the build platform specific crosstool-ng
+    # configuration values (e.g. CT_CONFIGURE_has_stat_flavor_GNU) get mixed with
+    # the host platform .config file.
+    yes "" | ct-ng oldconfig
+    ct-ng build
 fi
 
-# increase stack size to prevent test failures
-# http://gcc.gnu.org/bugzilla/show_bug.cgi?id=31827
-# ulimit -s 32768
+find . -name "*activate*.sh" -exec sed -i.bak "s|@CHOST@|${CHOST}|g" "{}" \;
+find . -name "*activate*.sh.bak" -exec sed rm "{}" \;
 
-# pushd .build/x86_64-sarc-linux-gnu/build/build-cc-gcc-final
-# make -k check || true
-# popd
-
-# .build/src/gcc-${PKG_VERSION}/contrib/test_summary
-
-chmod u+w -R ${PREFIX}
+chmod u+w -R ${SRC_DIR}/gcc_built
